Fix persistent maps attribute
Just looking for a key in a dict, which will always be there, isn't sufficient
to determine whether a robot has persistent maps or not. The persistent maps
endpoint seem to return empty lists if a robot doesn't have maps, and also the
string persistent_maps as a trait. Look for these to determine wheter a robot
has persistent maps or not.